requirements.txt
streamlit
yfinance
pandas
google-generativeai
python-dotenv
feedparser